+++ ledger/src/memdb.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{ledger::NUM_COLS, prelude::*};
+use crate::{prelude::*, NUM_COLS};
 
 use parking_lot::RwLock;
 use std::{collections::HashMap, path::Path};
